Made Auth Logic Similar to Exsisting Logic for Uniformity
