+++ .gitignore
@@ -5,3 +5,4 @@
 .ruff_cache
 *.sql
 /venv
+grep
